requirements.txt
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
matplotlib==3.9.3
seaborn==0.13.2
plotly==5.24.1
streamlit==1.40.2
Pillow==10.4.0
joblib==1.4.2
tqdm==4.66.4
opencv-python==4.10.0.84
